12 Pipeline与Valve：如何实现容器间的调用、事务管理、权限验证

+++ src/main/java/org/hermione/minit/Response.java
@@ -37,7 +37,7 @@ public interface Response {
 
     public void finishResponse() throws IOException;
 
-    public int getContentLength();
+    public long getContentLength();
 
     public String getContentType();
